Addition of tests for functionality; migration of gromacs stuff from utils to gromacs.
