+++ openmoltools/tests/test_gromacs.py
@@ -9,8 +9,8 @@ def test_gromacs_merge():
 
     with utils.enter_temp_directory(): #Prevents creating lots of tleap/antechamber files everywhere
         #Generate frcmod files, mol2 files
-        gaff_mol2_filename1, frcmod_filename1 = utils.run_antechamber( "etoh", etoh_filename, charge_method = None)
-        gaff_mol2_filename2, frcmod_filename2 = utils.run_antechamber( "benzene", benzene_filename, charge_method = None)
+        gaff_mol2_filename1, frcmod_filename1 = amber.run_antechamber( "etoh", etoh_filename, charge_method = None)
+        gaff_mol2_filename2, frcmod_filename2 = amber.run_antechamber( "benzene", benzene_filename, charge_method = None)
 
         #Set file names
         prmtop_filename1 = "./out1.prmtop"
@@ -23,15 +23,30 @@ def test_gromacs_merge():
         gro_filename2 = "./out2.gro"
 
         #Generate AMBER files
-        utils.run_tleap( 'etoh', gaff_mol2_filename1, frcmod_filename1, prmtop_filename1, crd_filename1 )
-        utils.run_tleap( 'benzene', gaff_mol2_filename2, frcmod_filename2, prmtop_filename2, crd_filename2 )
+        amber.run_tleap( 'etoh', gaff_mol2_filename1, frcmod_filename1, prmtop_filename1, crd_filename1 )
+        amber.run_tleap( 'benzene', gaff_mol2_filename2, frcmod_filename2, prmtop_filename2, crd_filename2 )
 
         #Convert to GROMACS
-        utils.convert_via_acpype( "etoh", prmtop_filename1, crd_filename1, out_top = top_filename1, out_gro = gro_filename1 ) 
-        utils.convert_via_acpype( "benzene", prmtop_filename2, crd_filename2, out_top = top_filename2, out_gro = gro_filename2 )
+        amber.convert_via_acpype( "etoh", prmtop_filename1, crd_filename1, out_top = top_filename1, out_gro = gro_filename1 ) 
+        amber.convert_via_acpype( "benzene", prmtop_filename2, crd_filename2, out_top = top_filename2, out_gro = gro_filename2 )
 
         #Merge topologies
         gromacs.merge_topologies( [ top_filename1, top_filename2], './combined.top', 'combined', molecule_numbers = [1, 5] )
 
         #Test editing of molecule numbers in topology file
         gromacs.change_molecules_section( './combined.top', './edited.top', ['etoh', 'benzene'], [10, 20] )
+
+
+def test_gromacs_solvate():
+    etoh_filename = utils.get_data_filename("chemicals/etoh/etoh.mol2")
+    with utils.enter_temp_directory(): #Prevents creating lots of tleap/antechamber files everywhere
+        #Generate frcmod files, mol2 files
+        gaff_mol2_filename, frcmod_filename = amber.run_antechamber( "etoh", etoh_filename, charge_method = None)
+
+        #Amber setup
+        amber.run_tleap( 'etoh', gaff_mol2_filename, frcmod_filename, 'etoh.prmtop', 'etoh.crd' )
+        #GROMACS conversion
+        utils.convert_via_acpype( 'etoh', 'etoh.prmtop', 'etoh.crd', 'etoh.top', 'etoh.gro' )
+        #Solvate
+        gromacs.do_solvate( 'etoh.top', 'etoh.gro', 'etoh_solvated.top', 'etoh_solvated.gro', 1.2, 'dodecahedron', 'spc216', 'tip3p.itp' )
+
