+++ openmoltools/utils.py
@@ -479,137 +479,3 @@ def get_checkmol_descriptors( molecule_filename, executable_name = 'checkmol' ):
  
     return descriptors
 
-def ensure_forcefield( intop, outtop, FF = 'ffamber99sb-ildn.ff'):
-    """Open a topology file, and check to ensure that includes the desired forcefield itp file. If not, remove any [ defaults ] section (which will be provided by the FF) and include the forcefield itp. Useful when working with files set up by acpypi -- these need to have a water model included in order to work, and most water models require a force field included in order for them to work.
-        
-        ARGUMENTS:
-        - intop: Input topology
-        - outtop: Output topology
-        OPTIONAL:
-        - FF: String corresponding to desired force field; default ffamber99sb.-ildn.ff
-        
-        Limitations:
-        - If you use this on a topology file that already includes a DIFFERENT forcefield, the result will be a topology file including two forcefields.
-        """
-    
-    file = open(intop, 'r')
-    text= file.readlines()
-    file.close()
-    
-    FFstring = FF+'/forcefield.itp'
-    
-    #Check if force field is included somewhere
-    found = False
-    for line in text:
-        if FFstring in line:
-            found = True
-    #If not, add it after any comments at the top
-    if not found:
-        idx = 0
-        while text[idx].find(';')==0:
-            idx+=1
-        text[idx] = '\n#include "%s"\n\n' % FFstring + text[idx]
-    
-    #Remove any defaults section
-    found = False
-    foundidx = None
-    endidx = None
-    for (idx, line) in enumerate(text):
-        if '[ defaults ]' in line:
-            foundidx = idx
-            found = True
-        #If we've already found the defaults section, find location of start of next section
-        #Assumes next section can be found by looking for a bracket at the start of a line
-        elif found and '[' in line:
-            #Only store if we didn't already store
-            if endidx == None:
-                endidx = idx
-    #Now remove defaults section
-    
-    if found:
-        text = text[0:foundidx] + text[endidx:]
-    
-    
-    #Write results
-    file = open( outtop, 'w')
-    file.writelines(text)
-    file.close()
-
-def do_solvate( top_filename, gro_filename, top_solv_filename, gro_solv_filename, box_dim, box_type, water_model, water_top, FF = 'amber99sb-ildn.ff' ):
-
-    """ This function creates water solvated molecule coordinate files and its corresponding topology
-        
-        PARAMETERS:
-            top_filename: str
-                          Topology path/filename
-            gro_filename: str
-                          Coordinates path/filename
-            top_solv_filename: str
-                          Topology path/filename (solvated system)
-            gro_solv_filename: str
-                          Coordinates path/filename (solvated system)
-            box_dim: float
-                          cubic box dimension (nm); will be passed to GROMACS with one digit of precision (%3.1f)
-            box_type: str
-                          box type (string passed to gmx solvate)
-            water_model: str
-                          Water model string to tell gmx solvate to use when solvating
-            water_top: str
-                          Water include file to ensure is present in topology file, i.e. "tip3p.itp"
-            FF : str, optional, default = 'amber99sb-ildn.ff'
-                          String specifying base force field directory for include files (i.e. 'amber99sb-ildn.ff').  
-
-        NOTES:
-        -----
-        Primarily tested on 3 point water models. May need adjustment for other models.
-"""
-
-    #Setting up proper environment variable (avoid unnecessary GROMCAS backup files)
-    os.environ['GMX_MAXBACKUP'] = '-1'
-
-
-    #copies topology file to new directory
-    shutil.copyfile(top_filename, top_solv_filename) 
-
-    #string with the Gromacs 5.0.4 box generating commands
-    cmdbox = 'gmx editconf -f %s -o %s -c -d %3.1f -bt %s' % (gro_filename, gro_solv_filename, box_dim, box_type)
-    output = getoutput(cmdbox)
-    logger.debug(output)
-
-    #string with the Gromacs 5.0.4 solvation tool (it is not genbox anymore)
-    cmdsolv = 'gmx solvate -cp %s -cs %s -o %s -p %s' % (gro_solv_filename, water_model, gro_solv_filename, top_solv_filename)
-    output = getoutput(cmdsolv)
-    logger.debug(output)
-
-    #Insert Force Field specifications
-    ensure_forcefield( top_solv_filename, top_solv_filename, FF = FF, version = 'new')
-
-    #Insert line for water topology portion of the code
-    try:
-        file = open(top_solv_filename,'r')
-        text = file.readlines()
-        file.close()
-    except:
-        raise NameError('The file %s is missing' % top_solv_filename)
-
-    #Insert water model
-    wateritp = os.path.join(FF, water_top ) # e.g water_top = 'tip3p.itp'
-    index = 0
-    while '[ system ]' not in text[index]:
-        index += 1
-    text[index] = '#include "%s"\n\n' % wateritp + text[index]
-
-    #Write the topology file
-    try:
-        file = open(top_solv_filename,'w+')
-        file.writelines( text )
-        file.close()
-    except:
-        raise NameError('The file %s is missing' % top_solv_filename)
-
-    #Check if file exist and is not empty;
-    if os.stat( gro_solv_filename ) == 0 or os.stat( top_solv_filename ).st_size == 0:
-        raise(ValueError("Solvent insertion failed"))
-
-    return
-
